requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
easyocr==1.7.1
torch>=1.8.0
# paddleocr==2.7.0.3
# paddlepaddle==2.6.2
numpy==1.26.4
opencv-python==4.6.0.66
pydantic==1.10.13
requests==2.31.0
python-multipart==0.0.9
gunicorn==21.2.0
prometheus-client==0.19.0
pytest==7.4.4
httpx==0.26.0
scipy==1.11.4
pillow==10.2.0
loguru==0.7.2
docker==7.0.0
